targetShootGame.java
This is the main game in which player has to hit the circle continuously as it get smaller and smaller.
Includes main game functionality and background music and wallpaper.

TODO: Update "targetMiss" functions (both, one is needed to call locally if time runs out, and the second is needed for if user miss clicks background screen and not target). Update them as per instructions. I recommend making an alternative screen with 2 buttons, one for 'Play Again' and another for 'View High Scores' which would connect to High Scores java file. 

This should be very easy to implement, let me know if any help is needed, I can do it although I have already invested allot of time on this and need to catch up on studying for exams. 

If any clarification on code is needed feel free to reach out, I spent allot of time researching and I can explain everything I did if needed
- Maaz