AndroidManifest.xml
Note in the manifest I have used AppTheme.NoActionBar. (Top title bar) was taking up game screen space, and did not seem sensible to be on game. So for other individual layouts, if we do want a header just create a header in drawable folder and reuse the same header on whatever screens necessary.
Thanks, Maaz